% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf_internal.R
\name{active_players_in_area}
\alias{active_players_in_area}
\title{Extract active players within defined area}
\usage{
active_players_in_area(active_players, centerpoint, dist = 20000)
}
\arguments{
\item{active_players}{A tibble as created by all_active_players()}

\item{centerpoint}{A polygonal location point calculated by sf::st_sfc)}

\item{dist}{A numerical value for distance (meters) fo cover from centerpoint}
}
\value{
A tibble (selection from active_players argument)
}
\description{
Extract active players within defined area
}
