% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf.R
\name{all_active_players}
\alias{all_active_players}
\title{Get (currently) active players}
\usage{
all_active_players()
}
\value{
A tibble with zone data with polygonal locations and retrieval time stamp
}
\description{
Get (currently) active players
}
