% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf.R
\name{display_zones_and_active_players}
\alias{display_zones_and_active_players}
\title{Display zones and active players, in an area of choice, on a map}
\usage{
display_zones_and_active_players(
  zones_all,
  address,
  turfarea_radius = 20,
  show_zones = TRUE,
  show_players = TRUE
)
}
\arguments{
\item{address}{A character string or vector (e.g, city, village, street)}

\item{turfarea_radius}{A numerical values, radius (in kilometers) of area to cover}

\item{show_zones}{TRUE (shows) or FALSE (does not show) zones on the map}

\item{show_players}{TRUE (shows) or FALSE (does not show) active players on the map}
}
\value{
Displays zones and active players
}
\description{
#' @param all_zones A tibble with info on all turfgame zones as created by turf_allzones_data()
}
