% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf_geo_KA.R
\name{get_geo_tibble}
\alias{get_geo_tibble}
\title{(HIDDEN) Helper to get_geo_pos}
\usage{
get_geo_tibble(address)
}
\arguments{
\item{address}{A character string or character vector}
}
\value{
A tibble with geocode data for the address
}
\description{
(HIDDEN) Helper to get_geo_pos
}
