% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf.R
\name{turf_get_all_zones_data}
\alias{turf_get_all_zones_data}
\title{Get all zones data
Stores all zones data in R_user_dir.
Read chached data if it exists, updates only if needed ,
and not again within less than 32 from previous succesful download.}
\usage{
turf_get_all_zones_data(
  max_age_days = 30,
  force_update = FALSE,
  allow_stale = TRUE
)
}
\arguments{
\item{max_age_days}{Numerical value, user preference for maximum age of cached data to be used.
(Is adjusted to be at least 32 minutes)}

\item{force_update}{If TRUE; Update data if the turfgame API just allows it. Defaults to FALSE.}

\item{allow_stale}{If TRUE: Allow use of outdated data if renewal fails. Defaults to TRUE.}
}
\value{
A tibble with all zones data
}
\description{
Get all zones data
Stores all zones data in R_user_dir.
Read chached data if it exists, updates only if needed ,
and not again within less than 32 from previous succesful download.
}
