% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf.R
\name{turf_statistics}
\alias{turf_statistics}
\title{Get current game statistics}
\usage{
turf_statistics()
}
\value{
A list with statistics such as `totalUsers` and `usersOnline`.
}
\description{
Get current game statistics
}
\examples{
\dontrun{
turf_statistics()
}
}
