% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf_internal.R
\name{turf_zones_all}
\alias{turf_zones_all}
\title{Send a request for all ones data and return result as ta tibble
NOTE: retrieval of all zones data is allowed by the turfgame API MAX ONCE PER 30 MINUTES}
\usage{
turf_zones_all()
}
\value{
A tibble with information on all turfgame zones,
with added polygonal locations calculated using sf:: package
}
\description{
Send a request for all ones data and return result as ta tibble
NOTE: retrieval of all zones data is allowed by the turfgame API MAX ONCE PER 30 MINUTES
}
