% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones_all}
\alias{zones_all}
\title{All Turf zones}
\format{
An sf object with one row per Turf zone.
}
\source{
\url{https://api.turfgame.com/v5}
}
\usage{
zones_all
}
\description{
A sample dataset containing information about all Turf zones,
retrieved from the Turf API 2026-10-07. Only for test use purposes!
}
\keyword{datasets}
