% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf_geo_KA.R
\name{display_zones_and_active_players}
\alias{display_zones_and_active_players}
\title{Title}
\usage{
display_zones_and_active_players(
  address,
  all_zones,
  turfarea_radius = 20,
  show_zones = TRUE,
  show_players = TRUE
)
}
\arguments{
\item{address}{A character string or vector}

\item{all_zones}{A tibble with info on all Turfgame zones}

\item{turfarea_radius}{A number, radius of area to cover}

\item{show_zones}{TRUE or FALSE}

\item{show_players}{TRUE or FALSE}
}
\value{
Displays zones and active players
}
\description{
Title
}
