% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf_internal.R
\name{get_geo_pos}
\alias{get_geo_pos}
\title{Get latitude and longitude for an address}
\usage{
get_geo_pos(address, prio_addresstype = NA_character_)
}
\arguments{
\item{address}{An address haracter string or vector}

\item{prio_addresstype}{Optional string for prioritized address type}
}
\value{
A tibble with the name of the address,
and its latitude and longitude
}
\description{
Get latitude and longitude for an address
}
