% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf_internal.R
\name{get_sf_point_from_koord}
\alias{get_sf_point_from_koord}
\title{Get a geometric point from koordinates}
\usage{
get_sf_point_from_koord(lat_long_vec)
}
\arguments{
\item{lat_long_vec}{Numerical vector of size 1x2, latitude in pos 1, longitude in pos 2}
}
\value{
sfc class object
}
\description{
Get a geometric point from koordinates
}
