% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf_geo_KA.R
\name{map_positions}
\alias{map_positions}
\title{map positions}
\usage{
map_positions(
  area_data,
  players = NULL,
  show_zones = TRUE,
  show_players = !is.null(players),
  show_legend = TRUE
)
}
\arguments{
\item{area_data}{sf A table with zones}

\item{players}{sf A table with players, defaults to NULL}

\item{show_zones}{TRUE or FALSE defaults to TRUE}

\item{show_players}{TRUE or FALSE, defult depending on players being NULL or not}

\item{show_legend}{TRUE or FALSE, defaults to TRUE}
}
\value{
prints map with optional players and zones
}
\description{
map positions
}
