% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf.R
\name{turf_users}
\alias{turf_users}
\title{Get users by name}
\usage{
turf_users(names)
}
\arguments{
\item{names}{A character vector of user names.}
}
\value{
A data frame with one row per user that was found.
}
\description{
Get users by name
}
\examples{
\dontrun{
turf_users(c("fredrick", "ingrid"))
}
}
