% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf.R
\name{turf_zones}
\alias{turf_zones}
\title{Get zones by zones}
\usage{
turf_zones(names)
}
\arguments{
\item{names}{A character vector of zone names.}
}
\value{
A data frame with one row per zone that was found.
}
\description{
Get zones by zones
}
\examples{
\dontrun{
turf_zones(c("Kårallen", "Östergötland"))
}
}
