% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf_KA.R
\name{turf_zones_all}
\alias{turf_zones_all}
\title{Get all zones}
\usage{
turf_zones_all()
}
\value{
A tibble with info on all zones,
vector geometry data added using sf:: package
time of request saved in tibble
}
\description{
NOTE: maximum one call per 30 minutes
}
