% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf_geo_KA.R
\name{zones_around}
\alias{zones_around}
\title{Gets zones within a given radius distance from
a specified point}
\usage{
zones_around(center, km, zones)
}
\arguments{
\item{center}{A sf point for center point of area in focus}

\item{km}{A number, radius distance to cover in kilometers}

\item{zones}{A tibble with all zones data the center point}
}
\value{
A sf table
}
\description{
Gets zones within a given radius distance from
a specified point
}
